refactor: move car-related's to domain package
