+++ src/main/java/racingcar/domain/CarValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class CarValidator {
