+++ src/main/java/racingcar/usecase/MoveOrStopCarWithNumberUseCase.java
@@ -1,6 +1,6 @@
 package racingcar.usecase;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 
 public class MoveOrStopCarWithNumberUseCase implements MoveOrStopCarUseCase {
